+++ changelog.md
@@ -19,6 +19,7 @@
 - Updated the `package.json` field `repository` to conform to new npm requirements.
 - Updated GitHub Actions CI config:
   - Updated the tested Node.js versions to v18, v20.
+- Migrated to the ESLint v9 CLI and “flat” config.
 - Removed the Node.js CLI option `--unhandled-rejections=throw` in the package script `tests` as it’s now the default for all supported Node.js versions.
 - Omit unused catch bindings in the function `processRequest`.